Add day 17 part 2 solution.

+++ src/main/scala/day17/PuzzleInputParser.scala
@@ -14,7 +14,7 @@ Register C: $registerC
 
 Program: $program""" =>
         (
-          Registers(registerA.toInt, registerB.toInt, registerC.toInt),
+          Registers(BigInt(registerA), BigInt(registerB), BigInt(registerC)),
           program.splitBy(",").toVector.map(_.toThreeBit)
         )
     }
